// src/types/hooksTypes.ts
export interface IPalette {
  THEME: 'LIGHT' | 'DARK';
  THEME_COLOR: string;
  MAIN_COLOR_1: string;
  MAIN_COLOR_2: string;
  MAIN_COLOR_3: string;
  MAIN_COLOR_4: string;
  MAIN_COLOR_5: string;
  GREY_COLOR_1: string;
  GREY_COLOR_2: string;
  GREY_COLOR_3: string;
  INFO_COLOR_1: string;
  INFO_COLOR_2: string;
}